Organização das pastas e projeto de consigurações
