+++ src/API/Program.cs
@@ -1,91 +1,29 @@
-using API.Configuration;
 using Cadastro.API.Interfaces;
 using Cadastro.API.Services;
+using Cadastro.Configuracoes;
 using Cadastro.Data.Repositories;
-using Cadastro.Domain.Entities;
 using Cadastro.Domain.Interfaces;
 using Microsoft.AspNetCore.Authentication.JwtBearer;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Builder;
-using Microsoft.AspNetCore.Http;
-using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Hosting;
 using Microsoft.IdentityModel.Logging;
-using Microsoft.IdentityModel.Tokens;
 using Microsoft.OpenApi.Models;
-using Newtonsoft.Json;
-using RabbitMQ.Client;
-using System.Collections.Generic;
-using System.IdentityModel.Tokens.Jwt;
-using System.Security.Claims;
-using System.Text;
 using System.Text.Json.Serialization;
-using System.Threading.Tasks;
-
 
 
 string MyAllowSpecificOrigins = "_myAllowSpecificOrigins";
-
-List<ClaimsIdentity> FillToken(TokenPayload payload)
-{
-
-    var claims = new List<Claim>();
-    claims.Add(new Claim(ClaimTypes.GivenName, payload.given_name));
-    claims.Add(new Claim(ClaimTypes.Name, payload.name));
-    claims.Add(new Claim(ClaimTypes.Email, payload.email));
-    claims.Add(new Claim(ClaimTypes.Surname, payload.family_name));
-
-    foreach (var item in payload.group)
-    {
-        claims.Add(new Claim(ClaimTypes.Role, item));
-    }
-    foreach (var item in payload.realm_access.roles)
-    {
-        claims.Add(new Claim(ClaimTypes.Role, item));
-    }
-    foreach (var item in payload.resource_access.account.roles)
-    {
-        claims.Add(new Claim(ClaimTypes.Role, item));
-    }
-
-    var identity = new ClaimsIdentity(claims);
-    return new List<ClaimsIdentity> { identity };
-}
-
-void SetupRabbitMQ(IConnection connection)
-{
-    try
-    {
-
-        IModel model = connection.CreateModel();
-        var cadastrarResult = model.QueueDeclare("cadastrar", true, false, false);
-        var atualizarrResult = model.QueueDeclare("atualizar", true, false, false);
-        var notificarResult = model.QueueDeclare("notificar", true, false, false);
-        model.ExchangeDeclare("cadastro", ExchangeType.Topic, true);
-
-        model.ExchangeDeclare("evento", ExchangeType.Fanout, true);
-        model.QueueBind("cadastrar", "cadastro", "cadastrar");
-        model.QueueBind("atualizar", "cadastro", "atualizar");
-        model.QueueBind("notificar", "evento", "");
-    }
-    catch (System.Exception ex)
-    {
-        throw ex;
-    }
-
-
-}
 
 var builder = WebApplication.CreateBuilder(args);
 
 // Add services to the container.
 
-builder.Services.AddControllers().AddJsonOptions(opt =>
-{
-    opt.JsonSerializerOptions.NumberHandling = JsonNumberHandling.AllowReadingFromString;
-    opt.JsonSerializerOptions.UnknownTypeHandling = JsonUnknownTypeHandling.JsonElement;
-});
+builder.Services.AddControllers()
+                .AddJsonOptions(opt =>
+                                {
+                                    opt.JsonSerializerOptions.NumberHandling = JsonNumberHandling.AllowReadingFromString;
+                                    opt.JsonSerializerOptions.UnknownTypeHandling = JsonUnknownTypeHandling.JsonElement;
+                                });
 // Learn more about configuring Swagger/OpenAPI at https://aka.ms/aspnetcore/swashbuckle
 builder.Services.AddEndpointsApiExplorer();
 
@@ -94,55 +32,7 @@ IdentityModelEventSource.ShowPII = true;
 System.Net.ServicePointManager.ServerCertificateValidationCallback +=
                         (sender, certificate, chain, sslPolicyErrors) => true;
 
-var clientId = builder.Configuration.GetValue<string>("ClientId");
-var clientSecret = builder.Configuration.GetValue<string>("ClientSecret");
-var key = new SymmetricSecurityKey(Encoding.ASCII.GetBytes(clientSecret)) { KeyId = clientId };
-
-var complement = builder.Configuration.GetValue<string>("UrlComplement");
-var authUrl = $"{builder.Configuration.GetValue<string>("BaseAuthUrl")}{complement}";
-var audience = builder.Configuration.GetValue<string>("Audience");
-
-builder.Services.AddAuthorization(options =>
-{
-    options.AddPolicy("email", policy => policy.Requirements.Add(new HasScopeRequirement("email", authUrl)));
-});
-
-builder.Services.AddAuthentication(JwtBearerDefaults.AuthenticationScheme)
-.AddJwtBearer(o =>
-{
-    o.Authority = authUrl;
-    if (!string.IsNullOrEmpty(audience))
-        o.Audience = audience;
-    o.RequireHttpsMetadata = false;
-    o.SaveToken = true;
-    o.ClaimsIssuer = "acme.com";
-    //  o.TokenValidationParameters.IssuerSigningKey = key;
-    o.TokenValidationParameters.NameClaimType = ClaimTypes.NameIdentifier;
-    o.TokenValidationParameters.ValidateIssuerSigningKey = false;
-    o.TokenValidationParameters.ValidateIssuer = false;
-    o.TokenValidationParameters.ValidIssuer = "acme.com";
-    o.Events = new JwtBearerEvents
-    {
-        OnTokenValidated = context =>
-        {
-            var token = (JwtSecurityToken)context.SecurityToken;
-            var payload = JsonConvert.DeserializeObject<TokenPayload>(token.Payload.SerializeToJson());
-            context.Principal.AddIdentities(FillToken(payload));
-            return Task.CompletedTask;
-        },
-        OnAuthenticationFailed = context =>
-        {
-            context.NoResult();
-            context.Response.StatusCode = 401;
-            context.Response.ContentType = "text/plain";
-            if (builder.Environment.IsDevelopment())
-            {
-                return context.Response.WriteAsync(context.Exception.ToString());
-            }
-            return context.Response.WriteAsync("An error occured processing your authentication.");
-        }
-    };
-});
+builder.Services.AddAPICustomAuthorizationAuthenticationConfig(builder.Environment, builder.Configuration);
 
 builder.Services.AddSwaggerGen(c =>
 {
@@ -190,14 +80,7 @@ builder.Services.AddScoped<IFuncionarioReadRepository, FuncionarioRepository>();
 
 builder.Services.AddScoped<IFuncionarioAppService, FuncionarioAppService>();
 
-builder.Services.AddSingleton(sp =>
-{
-    ConnectionFactory factory = new ConnectionFactory();
-    factory.Uri = new System.Uri(builder.Configuration.GetValue<string>("rabbit"));
-    IConnection connection = factory.CreateConnection();
-    SetupRabbitMQ(connection);
-    return connection;
-});
+builder.Services.AddRabbitCustomConfiguration(builder.Configuration);
 
 builder.Services.AddLogging();
 
@@ -215,6 +98,7 @@ if (!app.Environment.EnvironmentName.ToUpper().Contains("PROD"))
 }
 
 app.UseSwagger();
+
 app.UseSwaggerUI(c =>
 {
     c.SwaggerEndpoint("/swagger/v1/swagger.json", $"Cadastro API - {app.Environment.EnvironmentName} V1");
@@ -225,8 +109,6 @@ app.UseHttpsRedirection();
 
 app.UseAuthentication();
 app.UseAuthorization();
-
-
 
 app.UseCors();
