+++ src/Cadastro.GRPC/Cadastro.GRPC.csproj
@@ -1,8 +1,8 @@
-<Project Sdk="Microsoft.NET.Sdk.Web">
+﻿<Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
     <TargetFramework>net6.0</TargetFramework>
-    <Nullable>enable</Nullable>
+    <Nullable>disable</Nullable>
     <ImplicitUsings>enable</ImplicitUsings>
     <UserSecretsId>5250b97a-7bf5-49be-9a66-ce81cd33689d</UserSecretsId>
     <DockerDefaultTargetOS>Linux</DockerDefaultTargetOS>
