+++ src/Cadastro.WorkerService/Cadastro.WorkerService.csproj
@@ -2,7 +2,7 @@
 
   <PropertyGroup>
     <TargetFramework>net6.0</TargetFramework>
-	  <Nullable>enable</Nullable>
+	  <Nullable>disable</Nullable>
     <UserSecretsId>dotnet-Cadastro.WorkerService-4EA84E5C-CD43-4190-8E1A-F31D0214D9C6</UserSecretsId>
     <DockerDefaultTargetOS>Linux</DockerDefaultTargetOS>
     <DockerfileContext>..\..</DockerfileContext>
@@ -19,6 +19,7 @@
   </ItemGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\Cadastro.Configuracoes\Cadastro.Configuracoes.csproj" />
     <ProjectReference Include="..\Data\Cadastro.Data.csproj" />
     <ProjectReference Include="..\Domain\Cadastro.Domain.csproj" />
   </ItemGroup>
