+++ src/MVC/Controllers/BaseController.cs
@@ -20,7 +20,7 @@ namespace MVC.Controllers
             {
                 AccessToken = await HttpContext.GetTokenAsync("access_token");
                 IdToken = await HttpContext.GetTokenAsync("id_token");
-                var expire = await HttpContext.GetTokenAsync("expires_at");
+               // var expire = await HttpContext.GetTokenAsync("expires_at");
             }
         }
     }
